+++ packages/api-background-tasks-ddb/package.json
@@ -15,8 +15,7 @@
   "license": "MIT",
   "dependencies": {
     "@webiny/background-tasks": "0.0.0",
-    "@webiny/feature": "0.0.0",
-    "@webiny/plugins": "0.0.0"
+    "@webiny/feature": "0.0.0"
   },
   "devDependencies": {
     "@webiny/build-tools": "0.0.0",
