+++ packages/api-background-tasks-ddb/src/index.ts
@@ -1,10 +1,1 @@
-import type { Plugin } from "@webiny/plugins/types.js";
-import {
-    createBackgroundTaskContext,
-    createBackgroundTaskGraphQL
-} from "@webiny/background-tasks/api";
-
-export const createBackgroundTasks = (): Plugin[] => {
-    return [...createBackgroundTaskContext(), ...createBackgroundTaskGraphQL()];
-};
 export { BackgroundTasksDdbFeature } from "./BackgroundTasksDdbFeature.js";
