+++ packages/api-background-tasks-os/package.json
@@ -16,7 +16,6 @@
   "dependencies": {
     "@webiny/api-elasticsearch-tasks": "0.0.0",
     "@webiny/api-headless-cms-es-tasks": "0.0.0",
-    "@webiny/background-tasks": "0.0.0",
     "@webiny/plugins": "0.0.0"
   },
   "devDependencies": {
