+++ packages/api-background-tasks-os/tsconfig.build.json
@@ -4,7 +4,6 @@
   "references": [
     { "path": "../api-elasticsearch-tasks/tsconfig.build.json" },
     { "path": "../api-headless-cms-es-tasks/tsconfig.build.json" },
-    { "path": "../background-tasks/tsconfig.build.json" },
     { "path": "../plugins/tsconfig.build.json" }
   ],
   "compilerOptions": {
@@ -18,8 +17,6 @@
       "@webiny/api-elasticsearch-tasks": ["../api-elasticsearch-tasks/src"],
       "@webiny/api-headless-cms-es-tasks/*": ["../api-headless-cms-es-tasks/src/*"],
       "@webiny/api-headless-cms-es-tasks": ["../api-headless-cms-es-tasks/src"],
-      "@webiny/background-tasks/*": ["../background-tasks/src/*"],
-      "@webiny/background-tasks": ["../background-tasks/src"],
       "@webiny/plugins/*": ["../plugins/src/*"],
       "@webiny/plugins": ["../plugins/src"]
     }
