+++ packages/api-background-tasks-os/tsconfig.json
@@ -4,7 +4,6 @@
   "references": [
     { "path": "../api-elasticsearch-tasks" },
     { "path": "../api-headless-cms-es-tasks" },
-    { "path": "../background-tasks" },
     { "path": "../plugins" }
   ],
   "compilerOptions": {
@@ -18,8 +17,6 @@
       "@webiny/api-elasticsearch-tasks": ["../api-elasticsearch-tasks/src"],
       "@webiny/api-headless-cms-es-tasks/*": ["../api-headless-cms-es-tasks/src/*"],
       "@webiny/api-headless-cms-es-tasks": ["../api-headless-cms-es-tasks/src"],
-      "@webiny/background-tasks/*": ["../background-tasks/src/*"],
-      "@webiny/background-tasks": ["../background-tasks/src"],
       "@webiny/plugins/*": ["../plugins/src/*"],
       "@webiny/plugins": ["../plugins/src"]
     }
