+++ packages/api-elasticsearch-tasks/__tests__/helpers/useHandler.ts
@@ -1,7 +1,7 @@
 import { useContextHandler } from "@webiny/testing";
 import type { UseContextHandlerParams } from "@webiny/testing";
 import { createTestOpenSearchClient } from "@webiny/api-opensearch/testing";
-import { createBackgroundTaskContext, TasksCrud } from "@webiny/background-tasks/api";
+import { TasksCrud } from "@webiny/background-tasks/api";
 import { createElasticsearchBackgroundTasks } from "~/index";
 import type { Context } from "~/types";
 
@@ -13,7 +13,6 @@ export const useHandler = <C extends Context = Context>(params: Params = {}) => 
         // OpenSearch core is already registered by useContextHandler's cmsStorage (DDB-OS preset);
         // registering it again here throws "OpenSearch core must not be loaded more than once".
         plugins: [
-            ...createBackgroundTaskContext(),
             createElasticsearchBackgroundTasks(),
             ...[params.plugins].flat(Infinity as 1).filter(Boolean)
         ]
