+++ packages/api-headless-cms-ddb-es/__tests__/context/useHandler.ts
@@ -1,10 +1,6 @@
 import { Container } from "@webiny/di";
 import { RequestContainer, RequestContextInitializer } from "@webiny/event-handler-core";
-import {
-    GraphQLContextEnhancer,
-    GraphQLContextualSchema,
-    registerLegacyPluginsViaGqlContextualSchema
-} from "@webiny/handler-graphql";
+import { GraphQLContextEnhancer, GraphQLContextualSchema } from "@webiny/handler-graphql";
 import { ApiCoreFeature, registerApiCoreStorageOperations } from "@webiny/api-core";
 import { HeadlessCmsFeature } from "@webiny/api-headless-cms";
 import { TenantContext } from "@webiny/api-core/features/tenancy/TenantContext/abstractions.js";
@@ -14,7 +10,7 @@ import { CreateTenantUseCase } from "@webiny/api-core/exports/api/tenancy.js";
 import { loadWcpLicense } from "@webiny/api-core/features/wcp/loadWcpLicense.js";
 import { createTestWcpLicense } from "@webiny/wcp/testing/createTestWcpLicense.js";
 import { RegisterExtensionPlugin } from "@webiny/handler";
-import { createBackgroundTaskContext, TasksCrud } from "@webiny/background-tasks/api";
+import { BackgroundTasksFeature, TasksCrud } from "@webiny/background-tasks/api";
 import { createTestOpenSearchClient } from "@webiny/api-opensearch/testing";
 import { getStorageOps } from "@webiny/project-utils/testing/environment/index.js";
 import type { HeadlessCmsStorageOperations } from "@webiny/api-headless-cms/types";
@@ -96,17 +92,9 @@ export const useHandler = <C extends CmsContext = CmsContext>(params: CreateHand
             }
         }
 
-        const bgPlugins = ([createBackgroundTaskContext()].flat(Infinity as 1) as any[]).filter(
-            Boolean
-        );
-        processLegacyPlugins(
-            container,
-            bgPlugins.filter(p => p instanceof RegisterExtensionPlugin)
-        );
-        registerLegacyPluginsViaGqlContextualSchema(
-            container,
-            bgPlugins.filter(p => !(p instanceof RegisterExtensionPlugin))
-        );
+        // Background tasks are DI-native — the feature registers models, TasksCrud, and the GraphQL
+        // contextual schema (built below alongside the other contextual schemas).
+        BackgroundTasksFeature.register(container);
 
         const tenantCtx = container.resolve(TenantContext);
         tenantCtx.setTenant({
