+++ packages/api-website-builder-scheduler/package.json
@@ -28,6 +28,7 @@
     "@webiny/api-opensearch": "0.0.0",
     "@webiny/background-tasks": "0.0.0",
     "@webiny/build-tools": "0.0.0",
+    "@webiny/di": "^1.0.2",
     "@webiny/project-utils": "0.0.0",
     "@webiny/testing": "0.0.0",
     "typescript": "6.0.3",
