+++ packages/background-tasks/package.json
@@ -19,6 +19,8 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@graphql-tools/merge": "^9.1.10",
+    "@graphql-tools/schema": "^10.0.34",
     "@webiny/admin-ui": "0.0.0",
     "@webiny/api": "0.0.0",
     "@webiny/api-core": "0.0.0",
