+++ packages/background-tasks/src/api/TasksCrud.ts
@@ -1,8 +1,13 @@
 import { createAbstraction } from "@webiny/feature/api";
-import type { ITasksContextCrudObject } from "~/api/types.js";
+import type { ITasksContextObject } from "~/api/types.js";
 
-export const TasksCrud = createAbstraction<ITasksContextCrudObject>("Tasks/TasksCrud");
+/**
+ * At runtime this is the full tasks facade (CRUD + definition + service methods), assembled and
+ * registered by BackgroundTasksFeature. Typed as the full `ITasksContextObject` so sibling calls
+ * (e.g. `getDefinition`, `trigger`) type-check.
+ */
+export const TasksCrud = createAbstraction<ITasksContextObject>("Tasks/TasksCrud");
 
 export namespace TasksCrud {
-    export type Interface = ITasksContextCrudObject;
+    export type Interface = ITasksContextObject;
 }
