+++ packages/background-tasks/src/api/crud/cleanupTaskSubtree.ts
@@ -1,4 +1,6 @@
-import type { Context, ITask } from "~/api/types.js";
+import type { Container } from "@webiny/di";
+import type { ITask } from "~/api/types.js";
+import { TasksCrud } from "~/api/TasksCrud.js";
 
 /**
  * Recursively deletes the task identified by `rootId`, all its descendants, and
@@ -6,16 +8,16 @@ import type { Context, ITask } from "~/api/types.js";
  * best-effort: per-record errors are logged and swallowed, the function never
  * throws.
  */
-export const createCleanupTaskSubtree = (context: Context) => {
+export const createCleanupTaskSubtree = (container: Container) => {
     const listChildren = async (parentId: string): Promise<ITask[]> => {
-        const { items } = await context.tasks.listTasks({
+        const { items } = await container.resolve(TasksCrud).listTasks({
             where: { parentId }
         });
         return items;
     };
 
     const collectSubtree = async (rootId: string): Promise<ITask[]> => {
-        const root = await context.tasks.getTask(rootId);
+        const root = await container.resolve(TasksCrud).getTask(rootId);
         if (!root) {
             return [];
         }
@@ -37,17 +39,17 @@ export const createCleanupTaskSubtree = (context: Context) => {
     };
 
     const deleteTaskLogs = async (task: ITask): Promise<void> => {
-        const definition = context.tasks.getDefinition(task.definitionId);
+        const definition = container.resolve(TasksCrud).getDefinition(task.definitionId);
         if (!definition || definition.databaseLogs !== true) {
             return;
         }
         try {
-            const { items } = await context.tasks.listLogs({
+            const { items } = await container.resolve(TasksCrud).listLogs({
                 where: { task: task.id }
             });
             for (const log of items) {
                 try {
-                    await context.tasks.deleteLog(log.id);
+                    await container.resolve(TasksCrud).deleteLog(log.id);
                 } catch (ex: any) {
                     console.warn(
                         `cleanupTaskSubtree: failed to delete log "${log.id}" for task "${task.id}": ${ex.message}`
@@ -66,7 +68,7 @@ export const createCleanupTaskSubtree = (context: Context) => {
         for (const task of ordered) {
             await deleteTaskLogs(task);
             try {
-                await context.tasks.deleteTask(task.id);
+                await container.resolve(TasksCrud).deleteTask(task.id);
             } catch (ex: any) {
                 console.warn(
                     `cleanupTaskSubtree: failed to delete task "${task.id}": ${ex.message}`
