+++ packages/background-tasks/src/api/index.ts
@@ -1,11 +1,10 @@
 export type { Timer } from "./abstractions/Timer.js";
 export { TaskService } from "./domain/TaskService.js";
-export { createBackgroundTaskGraphQL } from "./graphql/index.js";
-export { createBackgroundTaskContext } from "./context.js";
 
 export * from "./response/index.js";
 export * from "./types.js";
 export { BackgroundTasksFeature } from "./BackgroundTasksFeature.js";
+export { BackgroundTasksContextualSchema } from "./graphql/BackgroundTasksContextualSchema.js";
 
 export { TasksCrud } from "./TasksCrud.js";
 export { TriggerTaskUseCase } from "./features/TriggerTask/abstractions.js";
